+++ Movecraft/src/main/java/net/countercraft/movecraft/features/contacts/ContactsCommand.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.commands;
+package net.countercraft.movecraft.features.contacts;
 
 import net.countercraft.movecraft.MovecraftLocation;
 import net.countercraft.movecraft.craft.Craft;
@@ -14,21 +14,20 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import static net.countercraft.movecraft.util.ChatUtils.MOVECRAFT_COMMAND_PREFIX;
 
 public class ContactsCommand implements CommandExecutor {
-
     @Override
-    public boolean onCommand(CommandSender commandSender, Command command, String s, String[] args) {
-        if (!command.getName().equalsIgnoreCase("contacts")) {
+    public boolean onCommand(@NotNull CommandSender commandSender, Command command, @NotNull String s, String[] args) {
+        if (!command.getName().equalsIgnoreCase("contacts"))
             return false;
-        }
-        if(!(commandSender instanceof Player)){
+
+        if(!(commandSender instanceof Player player)){
             commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Contacts - Must Be Player"));
             return true;
         }
-        Player player = (Player) commandSender;
 
         if (CraftManager.getInstance().getCraftByPlayer(player) == null) {
             player.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("You must be piloting a craft"));
@@ -37,19 +36,22 @@ public class ContactsCommand implements CommandExecutor {
 
         int page;
         try {
-            if (args.length == 0)
+            if (args.length == 0) {
                 page = 1;
-            else
+            }
+            else {
                 page = Integer.parseInt(args[0]);
-        }catch(NumberFormatException e){
+            }
+        } catch(NumberFormatException e){
             commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Paginator - Invalid Page") +" \"" + args[0] + "\"");
             return true;
         }
 
-        TopicPaginator pageinator = new TopicPaginator(I18nSupport.getInternationalisedString("Contacts"));
+        TopicPaginator paginator = new TopicPaginator(I18nSupport.getInternationalisedString("Contacts"));
         Craft ccraft = CraftManager.getInstance().getCraftByPlayer(player);
         HitBox hitBox = ccraft.getHitBox();
         MovecraftLocation center = hitBox.getMidPoint();
+        ContactsManager.update(ccraft);
         for (Craft tcraft : ccraft.getContacts()) {
             HitBox tHitBox = tcraft.getHitBox();
             if (tHitBox.isEmpty())
@@ -83,18 +85,18 @@ public class ContactsCommand implements CommandExecutor {
                 notification += " "+I18nSupport.getInternationalisedString("Contact/Subcraft Rotate - South") + ".";
             else
                 notification += " "+I18nSupport.getInternationalisedString("Contact/Subcraft Rotate - North") + ".";
-            pageinator.addLine(notification);
+            paginator.addLine(notification);
 
         }
-        if (pageinator.isEmpty()) {
+        if (paginator.isEmpty()) {
             player.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Contacts - None Found"));
             return true;
         }
-        if(!pageinator.isInBounds(page)){
+        if(!paginator.isInBounds(page)){
             commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Paginator - Invalid page") + "\"" + page + "\"");
             return true;
         }
-        for(String line :pageinator.getPage(page))
+        for(String line :paginator.getPage(page))
             commandSender.sendMessage(line);
         return true;
     }
